Fix building on Windows.
This fixes building on Windows. We disable reordering and fix iridium to link correctly on Chromium 55. We still download syzygy for now until we decide if we want to keep it. icu_use_data_file has to be enabled on Windows for now to get it to at least build. Patches are not read into memory and piped over stdin anymore.