devutils/validate_patches.py: Auto-determine GN version
Because there seems to be no reference to the logic for generating the
chromium-browser-official tar file, it's possible that it is being generated
by an internal script that manually injects the GN repository files.
Therefore, assume that the GN version used in the chromium-browser-official tar
files correspond to the latest commit in the master branch of the GN repository
at the time of the tar file's generation. We can get an approximation for the
generation time by using the last modification date of the tar file on
Google's file server.