Add missing includes to flag-fingerprinting-canvas-webgl-noise
This solution is a compromise/hack. The ideal solution would be to
extend third_party/skia with a public function that can do the data
shuffling, but that would require separating the Skia-dependent
constants from the algorithm, which is also interleaved with
dependencies on //base (which Skia isn't).

However, given that Skia seems to insert "Sk" or include "skia"
somewhere in all its names (file and variable names) means that it
should be fine to add those to the includes of
//third_party/blink/renderer/platform without much risk of
introducing a name conflict bug.

The Skia includes/private directory is chosen for inclusion since
SkColorData.h (and dependencies) only depend on includes within
that directory, so this would have the smallest impact in terms of
namespace pollution.

Fixes #508