devutils/validate_patches.py: Attempt to fix obscure bug
There was a CI job that failed due to being unable to find the fallback
repo for tools/gn/build/gen.py, even though the commit did not change
the Chromium source tree or any patches config. The previous and
following commits also did not have any error.

After a through review, I could not identify any logical errors in the
code. This is a far-fetched guess, but perhaps there is some hash
instability in Python 3.5?

See https://cirrus-ci.com/task/5744818110070784 for the failing CI job
in question.