utils/downloads.py: Remove axel downloader
When downloading the official Chromium tar file, axel is no longer able
to parallelize the connections nor detect the filesize. It is likely
Google added a specific rule for axel since curl and Chromium work
normally. While it is possible to circumvent this rule, Google might
have imposed it to reduce spikes of heavy load on their servers. We
should respect this and just use curl instead, which only uses one
connection and has the ability to resume interrupted downloads.