Add write permission to read-only files so their domains can be subsituted
This was added because both `third_party/shaka-player/dist/package.json` and
`third_party/shaka-player/dist/shaka-player.ui.extens.js` are read-only when
extracted from the chromium source tarball. Currently, this change could
potentially be replaced by manual vetting and whitelisting of the
beforementioned files from the domain subsitution routine. Another alternative
to this approach would be to pass an argument to tar (or whatever other program
is used to perform the extraction) that specifies that the extracted files are
to be writable.