Fix compiling with ThinLTO optimizations via the use_thin_lto=true flag
For some reason, building with ThinLTO (via the use_thin_lto=true flag)
fails when ungoogled-chromium patches are applied.

The failure is caused by undefined references to various crashpad
and breakpad symbols when linking.

It seems that building some of the crashpad/breakpad libraries without
LTO fixes this issues.

This adds two patches:
- the first allows for -Wl,--lto-O2 optimizations instead of just
  -Wl,--lto-O0 even if not official building
- the second separates the ThinLTO stuff from config("compiler") into
  config("thinlto"), adds this new config into default_compiler_configs,
  and then removes this config from the relevant crashpad and breakpad
  libraries, so that these libraries are build without LTO