Do not fail on fraction 49/49/1 condition
This ensures that when the A/B testing is done in 49/49/1 split, we do not
error out because the releases array is empty. Instead we only will error
out if the server returned >=300 response and/or returned invalid JSON (i.e.
a JSON without a releases array key).

Addresses https://github.com/ungoogled-software/ungoogled-chromium/pull/2555#issuecomment-1763397101
concern.

Signed-off-by: rany <ranygh@riseup.net>