Updated, rearranged software repositories and package managers
I added Slackware to the list of software repositories, and moved NixOS and GNU Guix to that category since their package managers are independent and closely related to their base distributions. I also added APK (for Alpine Linux and FreeBSD) and APT-RPM (for PCLinuxOS, Vine Linux, and ALT Linux) to the list of package managers.