Enable blocking all downloads via managed policy
Ungoogled chromium patches out the ability to block all
downloads via managed policy [1] as part of disabling
safebrowsing [2].

Only the handling of the "DownloadRestrictions" values
is related to safebrowsing, specifically the way that
key values 1, 2 and 4 are processed.

So let's modify things to treat these as invalid out
of range values (the same as the default case) which
emit a non-fatal error and allow downloads to proceed.

Values 0 and 3, which respectively enable and disable
all downloads, will work as expected.

[1] https://chromeenterprise.google/policies/#DownloadRestrictions
[2] core/ungoogled-chromium/fix-building-without-safebrowsing.patch

Fixes: https://github.com/ungoogled-software/ungoogled-chromium/issues/2974