depot_tools: Patch gclient.bat to call system python3 instead
Python is already required when building on Windows anyway, so we should
be using that instead of Google's "vpython".

This also resolves an issue where building would fail because httplib2
is not detected. This might have something to do with WSL allowing Linux
executables or shell scripts to be executed from cmd.